implemented, needs to be tested

+++ Test/SystemGameObjectQueue.cpp
@@ -57,7 +57,7 @@ void SystemGameObjectQueue::DeleteObjects(int g_id){
 	//}
 }
 
-GameObject* SystemGameObjectQueue::GetGameObject(int g_id) {
+GameObject* SystemGameObjectQueue::GetGameObject(uint64_t g_id) {
 		// iterate through alive objects 
 		for (auto iter = alive_objects.begin(); iter != alive_objects.end(); ++iter) {
 		if ((*iter)->ID == g_id) {
